feat: new contract deployment which whitelists bot for rebalancing
